# BEAST-Server/res/content/api-docs/OpenAPI.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "D-FARM API",
    "description" : "D-FARM API",
    "contact" : {
      "name" : "CodeInspect Support Team",
      "url" : "https://codeinspect.de",
      "email" : "helpdesk@codeinspect.de"
    },
    "version" : "0.2"
  },
  "security" : [ {
    "APIKey" : [ ]
  } ],
...
  "paths" : {
...
    "/api/devices/{devid}/input/android/broadcast" : {
      "post" : {
        "tags" : [ "Input" ],
        "summary" : "Sends an intent to the system to send a message to broadcast receivers",
        "description" : "Sends an intent to the system to send a message to broadcast receivers. Android only",
        "operationId" : "broadcast",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "receiverPermission",
          "in" : "query",
          "description" : "The permission a receiver needs to have",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "The intent ot send",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Intent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not an android device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "510" : {
            "description" : "No device available",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceReservationFailedException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/input/keyevent" : {
      "get" : {
        "tags" : [ "Input" ],
        "summary" : "Key event",
        "description" : "A key event",
        "operationId" : "keyTyped",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "key",
          "in" : "query",
          "description" : "The key",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "KEYCODE_SOFT_RIGHT", "KEYCODE_HOME", "KEYCODE_BACK", "KEYCODE_CALL", "KEYCODE_ENDCALL", "KEYCODE_0", "KEYCODE_1", "KEYCODE_2", "KEYCODE_3", "KEYCODE_4", "KEYCODE_5", "KEYCODE_6", "KEYCODE_7", "KEYCODE_8", "KEYCODE_9", "KEYCODE_STAR", "KEYCODE_POUND", "KEYCODE_DPAD_UP", "KEYCODE_DPAD_DOWN", "KEYCODE_DPAD_LEFT", "KEYCODE_DPAD_RIGHT", "KEYCODE_DPAD_CENTER", "KEYCODE_VOLUME_UP", "KEYCODE_VOLUME_DOWN", "KEYCODE_POWER", "KEYCODE_CAMERA", "KEYCODE_CLEAR", "KEYCODE_A", "KEYCODE_B", "KEYCODE_C", "KEYCODE_D", "KEYCODE_E", "KEYCODE_F", "KEYCODE_G", "KEYCODE_H", "KEYCODE_I", "KEYCODE_J", "KEYCODE_K", "KEYCODE_L", "KEYCODE_M", "KEYCODE_N", "KEYCODE_O", "KEYCODE_P", "KEYCODE_Q", "KEYCODE_R", "KEYCODE_S", "KEYCODE_T", "KEYCODE_U", "KEYCODE_V", "KEYCODE_W", "KEYCODE_X", "KEYCODE_Y", "KEYCODE_Z", "KEYCODE_COMMA", "KEYCODE_PERIOD", "KEYCODE_ALT_LEFT", "KEYCODE_ALT_RIGHT", "KEYCODE_SHIFT_LEFT", "KEYCODE_SHIFT_RIGHT", "KEYCODE_TAB", "KEYCODE_SPACE", "KEYCODE_SYM", "KEYCODE_EXPLORER", "KEYCODE_ENVELOPE", "KEYCODE_ENTER", "KEYCODE_DEL", "KEYCODE_GRAVE", "KEYCODE_MINUS", "KEYCODE_EQUALS", "KEYCODE_LEFT_BRACKET", "KEYCODE_RIGHT_BRACKET", "KEYCODE_BACKSLASH", "KEYCODE_SEMICOLON", "KEYCODE_APOSTROPHE", "KEYCODE_SLASH", "KEYCODE_AT", "KEYCODE_NUM", "KEYCODE_HEADSETHOOK", "KEYCODE_FOCUS", "KEYCODE_PLUS", "KEYCODE_MENU", "KEYCODE_NOTIFICATION", "KEYCODE_SEARCH", "KEYCODE_WAKEUP" ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "403" : {
            "description" : "Access denied",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AccessDeniedException"
                }
              }
            }
          },
          "404" : {
            "description" : "Device not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceNotFoundException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/input/android/startActivity" : {
      "post" : {
        "tags" : [ "Input" ],
        "summary" : "Sends an intent to the system to start an activity",
        "description" : "Sends an intent to the system to start an activity. Android only",
        "operationId" : "startActivity",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "forceStopBefore",
          "in" : "query",
          "description" : "Whether to force stop the application before",
          "schema" : {
            "type" : "boolean"
          },
          "example" : false
        }, {
          "name" : "waitForDebugger",
          "in" : "query",
          "description" : "Whether to wait for a debugger",
          "schema" : {
            "type" : "boolean"
          },
          "example" : false
        } ],
        "requestBody" : {
          "description" : "The intent ot send",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Intent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not an android device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "510" : {
            "description" : "No device available",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceReservationFailedException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/input/android/startForegroundService" : {
      "post" : {
        "tags" : [ "Input" ],
        "summary" : "Sends an intent to the system to start a service in foreground",
        "description" : "Sends an intent to the system to start a service in foreground. Android only",
        "operationId" : "startForegroundService",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "The intent ot send",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Intent"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not an android device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "510" : {
            "description" : "No device available",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceReservationFailedException"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/{devid}/input/android/startMonkey" : {
      "post" : {
        "tags" : [ "Input" ],
        "summary" : "Starts the monkey application exerciser",
        "description" : "Starts the monkey application exerciser, which randomly executes inputs on the device. Android only",
        "operationId" : "startMonkey",
        "parameters" : [ {
          "name" : "devid",
          "in" : "path",
          "description" : "The id of device",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "requestBody" : {
          "description" : "The options of monkey",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MonkeyOptions"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            }
          },
          "401" : {
            "description" : "Not an android device",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal error",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/APIException"
                }
              }
            }
          },
          "510" : {
            "description" : "No device available",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/DeviceReservationFailedException"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "APIException" : {
        "required" : [ "description", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short description of the error that occured",
            "example" : "Internal Error"
          },
          "description" : {
            "type" : "string",
            "description" : "A more detailed description of the error that occured",
            "example" : "Out of storage space"
          },
          "code" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "description" : "An exception has occurred while processing the request."
      },
...
      "DeviceReservationFailedException" : {
        "required" : [ "description", "title" ],
        "type" : "object",
        "properties" : {
          "title" : {
            "type" : "string",
            "description" : "A short description of the error that occured",
            "example" : "Internal Error"
          },
          "description" : {
            "type" : "string",
            "description" : "A more detailed description of the error that occured",
            "example" : "Out of storage space"
          },
          "code" : {
            "type" : "integer",
            "format" : "int32"
          }
        },
        "description" : "No device could be reserved"
      },
      "Intent" : {
        "required" : [ "action" ],
        "type" : "object",
        "properties" : {
          "action" : {
            "type" : "string",
            "description" : "The action",
            "example" : "android.intent.action.MAIN"
          },
          "category" : {
            "type" : "string",
            "description" : "The category",
            "example" : "android.intent.category.LAUNCHER"
          },
          "extras" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "string",
              "description" : "The extras"
            },
            "description" : "The extras"
          },
          "component" : {
            "type" : "string",
            "description" : "The component"
          },
          "dataURI" : {
            "type" : "string",
            "description" : "The data uri"
          },
          "mimeType" : {
            "type" : "string",
            "description" : "The mime type"
          },
          "flags" : {
            "type" : "integer",
            "description" : "Flags",
            "format" : "int32",
            "example" : 0
          }
        },
        "description" : "An Android Intent"
      },
...
    },
    "securitySchemes" : {
      "APIKey" : {
        "type" : "apiKey",
        "description" : "Your API key",
        "name" : "APIKey",
        "in" : "header"
      }
    }
  }
}